Remove condition managers interfaces
